add regex module, Missing to import
